+++ src/preset/bun-handler/index.ts
@@ -6,8 +6,13 @@ import { provideAbortController } from '../isomorphic/providers/abort-controller
 import { provideFetch } from '../isomorphic/providers/fetch';
 import { provideReduxMiddlewareSaga } from '../isomorphic/providers/redux-middleware-saga';
 import { providePageRender } from '../node-handler/providers/page-render';
+import { provideFetchLogHandler } from './providers/fetch-log-handler';
+import { provideFetchMiddleware } from './providers/fetch-middleware';
+import { provideHandlerMain } from './providers/handler-main';
+import { providePageHelmet } from './providers/page-helmet';
+import { provideSpecificParams } from './providers/specific-params';
+import { provideCookieStore } from './providers/cookie-store';
 import { SpecificExtras } from '../server/utils/specific-extras';
-import { HandlerProviders } from './providers';
 
 /**
  * Возвращает preset с зависимостями для формирования обработчика входящего http-запроса.
@@ -20,16 +25,16 @@ export function PresetBunHandler(customize?: PresetTuner) {
   // http fetch
   preset.set(KnownToken.Http.fetch, provideFetch);
   preset.set(KnownToken.Http.Fetch.abortController, provideAbortController);
-  preset.set(KnownToken.Http.Fetch.cookieStore, HandlerProviders.cookieStore);
-  preset.set(KnownToken.Http.Fetch.middleware, HandlerProviders.fetchMiddleware);
-  preset.set(KnownToken.Http.Fetch.Middleware.Log.handler, HandlerProviders.fetchLogHandler);
+  preset.set(KnownToken.Http.Fetch.cookieStore, provideCookieStore);
+  preset.set(KnownToken.Http.Fetch.middleware, provideFetchMiddleware);
+  preset.set(KnownToken.Http.Fetch.Middleware.Log.handler, provideFetchLogHandler);
 
   // handler
-  preset.set(KnownToken.Http.Handler.main, HandlerProviders.handlerMain);
-  preset.set(KnownToken.Http.Handler.Request.specificParams, HandlerProviders.specificParams);
+  preset.set(KnownToken.Http.Handler.main, provideHandlerMain);
+  preset.set(KnownToken.Http.Handler.Request.specificParams, provideSpecificParams);
   preset.set(KnownToken.Http.Handler.Response.specificExtras, () => new SpecificExtras());
   preset.set(KnownToken.Http.Handler.Page.assets, () => ({ js: '', css: '' }));
-  preset.set(KnownToken.Http.Handler.Page.helmet, HandlerProviders.pageHelmet);
+  preset.set(KnownToken.Http.Handler.Page.helmet, providePageHelmet);
   preset.set(KnownToken.Http.Handler.Page.render, providePageRender);
 
   // redux saga
